RL_Book/requirements.txt
pyyaml>=6.0.3
gymnasium>=0.26.0
gymnasium[box2d]>=0.26.0
pyglet>=2.0
pygame>=2.0
tensorboard>=2.0
ray[default]>=2.0
numpy>=1.21

# ONNX 내보내기 (export_onnx.py)
onnx>=1.21.0
onnxruntime>=1.24.0
onnxscript>=0.6.0